requirements.txt
pydantic
telethon
PyYAML
pydantic_settings
